refactor(server): extract reusable git-checkout/materialization core
Lift the git-clone + manifest-from-checkout machinery out of
automation_materializer into a new `git_checkout` module so it can be
reused beyond automations: GitRepoCache (cached bare clone + worktree),
the git command plans, credential resolution/redaction, and GitHub
owner/repo slug parsing/validation.

Split build_manifest_from_checkout so the workflow-source checkout (which
workflow to bundle) and the git context (which repo the run clones and
executes in) are separate inputs. Automations are now the case where both
point at the same repo; this is the seam a future external-workflow
resolver needs to bundle a workflow from one repo and run it against
another.

ManifestFromCheckoutInput no longer embeds AutomationRunMaterializeInput;
it takes only the fields it needs plus a caller-supplied error context, so
the builder is reusable without automation-specific types.

The shared materialize error moves into git_checkout as the
provider-neutral RunMaterializeError (same variants, neutral messages) so
the foundation module no longer depends back on its consumer.

The workflow git context is a required input rather than Option; no caller
omits it today, and it can widen to optional later without a breaking
change.

No behavior change for automations: the assembled manifest, git context,
and serialized bytes are identical.
